Make stan_csv_reader insensitive to whitespace around rhs of .csv file metadata.
Previously running stansummary on a .csv file produced by either rstan or pystan
would fail on a metadata line with no space after the "=" sign.  Editing the
files to say "= " would fix the problem.  With this change any amount of spacing
before or after the value works as it gets trimmed before the lexical_cast.